+++ changelog.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Fixed
+
+- Heroes now open the city destination list when using Transfer to Another City.
+
 ## [1.5.1] - 2026-09-24
 
 ### Fixed